Begin test program with tests for escape.
The exact escaping of of escape is unspecified; it says only that it
will encode so as to remove certain byte values. The idea is that you
shouldn't be able to inject a line by sending '\n', or truncate one with
'\0'. So the tests don't look for specific output values, only that they
don't contain any forbidden bytes.