Remove checks that at least one method name was selected.
I meant this that ClientSetup and ServerSetup would never return without
at least one method being selected, so that you could use successful
return of the function as a signal to go ahead and open your listener,
if you knew you were dealing with only one possible method name. But in
general you have to deal with all the method names that are returned (in
client and server mode equally), and now the library won't exit on you.
The calling application has to do somethign intelligent when it gets an
empty method list back.