Don't use this library.
I'm going to be making some incompatible changes, probably up until
support for transport args. What is in this branch is already
incompatible with what's in /pluggable-transports/websocket.git/src/pt
because of changed error handling. I'll update programs that use this
library when I change their import paths to point to this repo.