Fix TestResolveAddr.
We try reading the address from the Buffer with a size one greater than
its expected length, to ensure that it's not too long. But we didn't
trim that extra byte before trying to resolve it again, so there was an
extra '\x00' at the end. Previously it was ignored by net.LookupPort.