Use the argument to ServerSetup only when the "*" list is requested.
Formerly the list always served as an additional filter of methods that
we support. Now, ServerSetup always returns all the method names
requested by Tor, unless "*" is requested, in which case the given
method name list is returned.

This change is to enable the case where the list of method names is not
known in advance.

https://lists.torproject.org/pipermail/tor-dev/2013-December/005966.html