Use the argument to ClientSetup only when the "*" list is requested.
Formerly the list always served as an additional filter of methods
that we support. Now, ClientSetup always returns all the method names
requested by Tor, unless "*" is requested, in which case the given
method name list is returned.

This change is to enable the case where for "*" you want to return
something other than the full list of transports you support.

https://lists.torproject.org/pipermail/tor-dev/2013-December/005966.html